externalise computation of hist comparison (part 2)
